requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pyspark==4.0.1
pandas==2.3.3
numpy==2.3.3
python-dotenv==1.1.1
psycopg2-binary==2.9.10
langchain-groq==0.3.8
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-postgres==0.0.15
langchain-openai==0.3.35
azure-search-documents==11.6.0
azure-identity==1.25.1
langchain-text-splitters==0.3.11
python-multipart==0.0.20
python-docx==1.2.0
asyncio==4.0.0
docx2txt==0.9
pyodbc==5.2.0
langgraph==1.0.1